[#15]refactor: ClassroomAccount 폴더를 class.account로 이동
